modified bookID Integer to long

+++ book-service/src/main/java/bjit/ursa/bookservice/controller/BookController.java
@@ -19,12 +19,12 @@ public class BookController {
     }
 
     @DeleteMapping("/delete/{bookId}")
-    public ResponseEntity<Object> deleteBookById(@PathVariable Integer bookId) {
+    public ResponseEntity<Object> deleteBookById(@PathVariable long bookId) {
         return ResponseEntity.ok(bookService.deleteBookById(bookId));
     }
 
     @PutMapping("/update/bookId/{bookId}")
-    public ResponseEntity<Object> updateBook(@RequestParam Integer bookId , @RequestBody BookEntity bookEntity) {
+    public ResponseEntity<Object> updateBook(@RequestParam long bookId , @RequestBody BookEntity bookEntity) {
         return bookService.updateBooks(bookId , bookEntity);
     }
 
@@ -34,7 +34,7 @@ public class BookController {
     }
 
     @GetMapping("/id/{bookId}")
-    public ResponseEntity<Object> getBookById(@PathVariable Integer bookId) {
+    public ResponseEntity<Object> getBookById(@PathVariable long bookId) {
         return bookService.getBookById(bookId);
     }
 }
